Update database schema and enhance logging level
The main action in this commit is the movement of all database tables to the "movie" schema. Changes include updating the queries in DAOs, modifying the "script.sql" file to reflect this change, and including a new "SchemaCreatingIntegrator" for creating the schema if it doesn't exist. This commit also increases the logging level for Hibernate queries for better debugging and fixes an indentation issue in "Main.java".